% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2GRange.R
\name{convert2GRange}
\alias{convert2GRange}
\title{convert output of BigLD/GPART to "GRangesList" object}
\usage{
convert2GRange(blockresult)
}
\arguments{
\item{blockresult}{BigLD or GPART output}
}
\value{
\code{GRangeList} object including the BigLD or GPART output
}
\description{
\code{convert2GRange} convert a BigLD or GPART output to a data of \code{GRangesList} object.
}
\examples{

testBigLD <- BigLD(geno=geno[,1:100], SNPinfo=SNPinfo[1:100,])
testBigLD_grange <- convert2GRange(testBigLD)

}
\author{
Sun-Ah Kim <sunny03@snu.ac.kr>, Yun Joo Yoo <yyoo@snu.ac.kr>
}
