% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpart.R
\docType{data}
\name{geneinfo}
\alias{geneinfo}
\title{gene information data}
\format{A data frame with 736 rows and 4 columns for genename, chromosome name, start bp and end bp of each gene.}
\source{
<http://grch37.ensembl.org/index.html>
}
\usage{
data(geneinfo)
}
\description{
This data set gives gene information in chromosome 21.
}
\references{
{
Zerbino, Daniel R., et al. "Ensembl 2018." \emph{Nucleic acids research 46.D1} (2017): D754-D761.
}
}
\keyword{datasets}
