% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpart.R
\docType{data}
\name{geno}
\alias{geno}
\title{genotype data}
\format{A data frame with 286 rows and 9000 columns}
\source{
1000 genomes project Phase 1 dataset <http://www.internationalgenome.org/>
}
\usage{
data(geno)
}
\description{
This data set gives genotype data that are subset of 1000 Genomes Project phase1 release 3 genotype data
with 286 individuals from JPT, CHB and CHS populations (1000 Genomes Project Consortium, 2012).
The dataset contains 9000 SNPs in chromosome 21
}
\references{
{
 1000 Genomes Project Consortium.
 "An integrated map of genetic variation from 1,092 human genomes." \emph{Nature} 491.7422 (2012): 56.
 }
}
\keyword{datasets}
