% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpart-package.R
\docType{package}
\name{gpart-package}
\alias{gpart}
\alias{gpart-package}
\title{gpart: Human genome partitioning of dense sequencing data by identifying haplotype blocks}
\description{
we provide a new SNP sequence partitioning method which partitions the whole SNP sequence based on not only LD block structures but also gene location information. The LD block construction for GPART is performed using Big-LD algorithm, with additional improvement from previous version reported in Kim et al.(2017). We also add a visualization tool to show the LD heatmap with the information of LD block boundaries and gene locations in the package.
}
\author{
\strong{Maintainer}: Sun Ah Kim \email{sunnyeesl@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Yun Joo Yoo [copyright holder]
}

}
\keyword{internal}
