% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpart.R
\docType{data}
\name{SNPinfo}
\alias{SNPinfo}
\title{SNP information data}
\format{A data frame with 9000 rows and 3 columns for chromosome name, rsID and bp}
\source{
1000 genomes project Phase 1 dataset <http://www.internationalgenome.org/>
}
\usage{
data(SNPinfo)
}
\description{
This data set gives information data of SNPs in \code{geno}
The dataset contains chromosome name, rsID and bp information of the 9000 SNPs in \code{geno}
}
\references{
{
 1000 Genomes Project Consortium.
 "An integrated map of genetic variation from 1,092 human genomes." \emph{Nature} 491.7422 (2012): 56.
 }
 @keywords datasets
}
